requirements.txt
fastapi
uvicorn
pdfplumber
requests
google-generativeai
python-dotenv
sqlalchemy
psycopg2-binary
